pyproject.toml
[project]
name = "backgroundannotator"
version = "1.0"
description = "BOB: Brain On Background"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.0.0",
]

[project.scripts]
app = "backgroundannotator.app:main"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ['src/backgroundannotator']
